requirements.txt
alembic==1.10.2
anyio==3.6.2
astor==0.8.1
attrs==22.2.0
catboost==1.1.1
category-encoders==2.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
colour==0.1.5
contextlib2==21.6.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.33
dnspython==2.3.0
dtreeviz==2.2.0
exceptiongroup==1.1.1
filelock==3.10.0
fonttools==4.39.2
future==0.18.3
gensim==4.3.1
graphviz==0.20.1
greenlet==2.0.2
h11==0.14.0
hyperopt==0.1.2
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
json-tricks==3.16.1
kiwisolver==1.4.4
lightgbm==3.3.5
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.0
nni==2.9
numba==0.56.4
numpy==1.23.5
optuna==3.1.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
pip==23.0.1
plotly==5.13.1
pluggy==1.0.0
prettytable==3.6.0
psutil==5.9.4
pycparser==2.21
pydantic==1.10.6
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
responses==0.23.1
schedule==1.1.0
schema==0.7.5
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
shap==0.41.0
simplejson==3.18.4
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
statsmodels==0.13.5
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
typeguard==2.9.0
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
websockets==10.4
wheel==0.38.4
wordcloud==1.8.2.2
xgboost==1.7.4
